bin/requirements.txt
socketIO-client-2-master.zip